text-sentiment-server/requirements.txt
click==7.1.2
joblib==1.0.0
nltk==3.5
numpy==1.19.4
pandas==1.1.5
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
regex==2020.11.13
six==1.15.0
textblob==0.15.3
tqdm==4.54.1
vosk==0.3.15
websockets==8.1
